requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.0
Faker==1.0.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
six==1.12.0
text-unidecode==1.2
